Add frr.vtysh.pathspace option (=> vtysh -N)
If you have multiple instances of frr running on the same machine
(e.g. because they are responsible for different network namespaces)
this option is necessary to be able to export metrics from all frr
instances.

vtysh supports the -N option (they call it pathspace, the prefix for frr config)

This commit adds support for passing -N to vtysh