Adopt structured logging using log/slog
Prometheus and its associated libraries are moving away from go-kit/log,
and instead will use the built-in log/slog from the Go standard library,
which is available since Go 1.21.

prometheus/exporter-toolkit v0.13.0 implemented a
non-backwards-compatible change to the web.ListenAndServe() function
(among others), and now expects to be supplied a *slog.Logger.

For the end user, this change is minimal. Logs may have a subtly
different appearance, but are still in logfmt, so log parsers such as
Loki should not trip up on this.

Signed-off-by: Daniel Swarbrick <daniel.swarbrick@gmail.com>